requirements.txt
flask
flask-wtf
pyserial
flask-sqlalchemy
flask-migrate
flask_apscheduler
flask-socketio == 4.3.2
flask_mqtt
Flask-Cors
psycopg2-binary
dependency-injector
requests
